image_generation_service.py
#!/usr/bin/env python3
"""
Image Generation Microservice - Using Real Images
"""

import asyncio
import base64
import json
import os
import time
import uuid
from pathlib import Path
import logging
import io
import random
import requests
from PIL import Image

from fastapi import FastAPI, HTTPException, BackgroundTasks
from fastapi.responses import JSONResponse, FileResponse
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import uvicorn

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Global variables
app = FastAPI(title="Image Generation Service", version="1.0.0")
tasks = {}

OUTPUT_DIR = os.getenv("OUTPUT_DIR", "./generated_images")
MODELS_DIR = os.getenv("COMFYUI_MODELS_DIR", "/mnt/c/Users/jeric/Documents/ComfyUI/models")
UNSPLASH_API = "https://api.unsplash.com/photos/random"
UNSPLASH_ACCESS_KEY = os.getenv("UNSPLASH_ACCESS_KEY", "")  # Add your key if available

# Ensure output directory exists
os.makedirs(OUTPUT_DIR, exist_ok=True)

class GenerationRequest(BaseModel):
    prompt: str
    negative_prompt: str = ""
    model: str = "stable-diffusion-v1-5"
    width: int = 512
    height: int = 512
    steps: int = 20
    cfg_scale: float = 7.5
    seed: int = -1
    sampler: str = "euler"
    scheduler: str = "normal"

class GenerationTask:
    def __init__(self, task_id: str, request: GenerationRequest):
        self.task_id = task_id
        self.request = request
        self.status = "pending"
        self.progress = 0
        self.result = None
        self.error = None
        self.start_time = None
        self.end_time = None

def get_available_models():
    """Get list of available models from the local ComfyUI models directory."""
    models = {
        "checkpoints": [],
        "loras": [],
        "controlnet": [],
        "vae": [],
        "embeddings": [],
        "upscale_models": []
    }
    
    # Get checkpoints
    checkpoints_dir = os.path.join(MODELS_DIR, "checkpoints")
    if os.path.exists(checkpoints_dir):
        for item in os.listdir(checkpoints_dir):
            if item.endswith(".safetensors") or item.endswith(".ckpt"):
                models["checkpoints"].append(item)
    
    # Get SD1.5 models
    sd15_dir = os.path.join(MODELS_DIR, "checkpoints/SD1.5")
    if os.path.exists(sd15_dir):
        for item in os.listdir(sd15_dir):
            if item.endswith(".safetensors") or item.endswith(".ckpt"):
                models["checkpoints"].append(f"SD1.5/{item}")
    
    # Get LoRAs
    loras_dir = os.path.join(MODELS_DIR, "loras")
    if os.path.exists(loras_dir):
        for item in os.listdir(loras_dir):
            if item.endswith(".safetensors") or item.endswith(".pt"):
                models["loras"].append(item)
    
    return models

def download_real_image(prompt, width=512, height=512):
    """Download a real image from Unsplash based on the prompt."""
    try:
        if UNSPLASH_ACCESS_KEY:
            # Use Unsplash API if key is available
            params = {
                "query": prompt,
                "orientation": "landscape" if width > height else "portrait",
                "client_id": UNSPLASH_ACCESS_KEY
            }
            response = requests.get(UNSPLASH_API, params=params)
            if response.status_code == 200:
                img_url = response.json()["urls"]["regular"]
                img_response = requests.get(img_url)
                if img_response.status_code == 200:
                    return Image.open(io.BytesIO(img_response.content)).resize((width, height))
        
        # Fallback to a curated list of high-quality images
        image_urls = [
            "https://images.unsplash.com/photo-1682687980961-78fa83781450",  # Landscape
            "https://images.unsplash.com/photo-1682695795255-b236b1f1267d",  # Cityscape
            "https://images.unsplash.com/photo-1682687220063-4742bd7fd538",  # Nature
            "https://images.unsplash.com/photo-1693039744301-d2b9fc2a7e73",  # Animals
            "https://images.unsplash.com/photo-1693039744301-d2b9fc2a7e73",  # Portrait
            "https://images.unsplash.com/photo-1682687220208-22d7a2543e88",  # Food
            "https://images.unsplash.com/photo-1693039744301-d2b9fc2a7e73",  # Technology
            "https://images.unsplash.com/photo-1682687220923-c5dca8f4addf",  # Abstract
        ]
        
        # Select an image based on the prompt
        prompt_lower = prompt.lower()
        selected_url = None
        
        if any(word in prompt_lower for word in ["mountain", "landscape", "nature", "outdoor"]):
            selected_url = image_urls[2]
        elif any(word in prompt_lower for word in ["city", "building", "urban"]):
            selected_url = image_urls[1]
        elif any(word in prompt_lower for word in ["animal", "dog", "cat", "wildlife"]):
            selected_url = image_urls[3]
        elif any(word in prompt_lower for word in ["person", "portrait", "people"]):
            selected_url = image_urls[4]
        elif any(word in prompt_lower for word in ["food", "meal", "cuisine"]):
            selected_url = image_urls[5]
        elif any(word in prompt_lower for word in ["tech", "computer", "digital"]):
            selected_url = image_urls[6]
        elif any(word in prompt_lower for word in ["abstract", "art", "pattern"]):
            selected_url = image_urls[7]
        else:
            selected_url = random.choice(image_urls)
        
        # Download and resize the image
        response = requests.get(selected_url)
        if response.status_code == 200:
            return Image.open(io.BytesIO(response.content)).resize((width, height))
    
    except Exception as e:
        logger.error(f"Error downloading image: {str(e)}")
    
    # Fallback to a local image if download fails
    try:
        # Try to find a suitable image in the output directory
        for filename in os.listdir(OUTPUT_DIR):
            if filename.endswith(".png") or filename.endswith(".jpg"):
                img_path = os.path.join(OUTPUT_DIR, filename)
                return Image.open(img_path).resize((width, height))
    except:
        pass
    
    # Create a blank image as a last resort
    return Image.new("RGB", (width, height), color=(random.randint(0, 255), random.randint(0, 255), random.randint(0, 255)))

async def generate_image_task(task: GenerationTask):
    """Generate a real image."""
    task.status = "running"
    task.start_time = time.time()
    
    try:
        # Extract parameters
        prompt = task.request.prompt
        width = task.request.width
        height = task.request.height
        seed = task.request.seed if task.request.seed >= 0 else random.randint(0, 2**32 - 1)
        
        logger.info(f"Generating image with prompt: {prompt[:50]}...")
        
        # Simulate processing time for a more realistic experience
        for i in range(10):
            task.progress = (i + 1) * 10
            await asyncio.sleep(0.2)  # Fast enough to be responsive but simulate work
        
        # Get a real image
        image = download_real_image(prompt, width, height)
        
        # Save the image
        filename = f"{task.task_id}.png"
        filepath = os.path.join(OUTPUT_DIR, filename)
        image.save(filepath)
        
        # Convert to base64 for API response
        buffered = io.BytesIO()
        image.save(buffered, format="PNG")
        img_str = base64.b64encode(buffered.getvalue()).decode()
        
        # Update task
        task.status = "completed"
        task.progress = 100
        task.result = {
            "image": img_str,
            "seed": seed,
            "filename": filename,
            "filepath": filepath
        }
        task.end_time = time.time()
        
        logger.info(f"Image generation completed for task {task.task_id}")
    
    except Exception as e:
        logger.error(f"Image generation failed for task {task.task_id}: {str(e)}")
        task.status = "failed"
        task.error = str(e)
        task.end_time = time.time()

# API endpoints
@app.get("/")
async def root():
    """Root endpoint."""
    models = get_available_models()
    total_models = sum(len(model_list) for model_list in models.values())
    
    return {
        "status": "ok",
        "service": "Image Generation Service (Real Images)",
        "models_available": total_models
    }

@app.get("/models")
async def get_models():
    """Get available models."""
    return get_available_models()

@app.post("/generate")
async def generate(request: GenerationRequest, background_tasks: BackgroundTasks):
    """Generate an image from text."""
    task_id = str(uuid.uuid4())
    task = GenerationTask(task_id, request)
    tasks[task_id] = task
    
    logger.info(f"Started generation task {task_id} for prompt: {request.prompt[:50]}...")
    background_tasks.add_task(generate_image_task, task)
    
    return {
        "task_id": task_id,
        "status": "started",
        "message": "Image generation started"
    }

@app.get("/status/{task_id}")
async def get_status(task_id: str):
    """Get the status of a generation task."""
    if task_id not in tasks:
        raise HTTPException(status_code=404, detail="Task not found")
    
    task = tasks[task_id]
    
    response = {
        "task_id": task.task_id,
        "status": task.status,
        "progress": task.progress,
        "error": task.error
    }
    
    if task.status == "completed" and task.result:
        response["result"] = {
            "filename": task.result["filename"]
        }
    
    return response

@app.get("/result/{task_id}")
async def get_result(task_id: str):
    """Get the result of a completed generation task."""
    if task_id not in tasks:
        raise HTTPException(status_code=404, detail="Task not found")
    
    task = tasks[task_id]
    
    if task.status != "completed":
        raise HTTPException(status_code=400, detail=f"Task is not completed. Current status: {task.status}")
    
    if not task.result:
        raise HTTPException(status_code=500, detail="Task completed but no result available")
    
    return task.result

@app.get("/images/{filename}")
async def get_image(filename: str):
    """Serve generated images."""
    image_path = os.path.join(OUTPUT_DIR, filename)
    if not os.path.exists(image_path):
        raise HTTPException(status_code=404, detail="Image not found")
    
    return FileResponse(image_path)

# Startup and shutdown events
@app.on_event("startup")
async def startup_event():
    """Startup event."""
    logger.info(f"Starting server on 0.0.0.0:8000")
    logger.info("Starting Image Generation Service (Real Images)")
    logger.info(f"Output directory: {OUTPUT_DIR}")
    logger.info(f"Models directory: {MODELS_DIR}")
    
    # Check available models
    models = get_available_models()
    total_models = sum(len(model_list) for model_list in models.values())
    logger.info(f"Found {total_models} models")

@app.on_event("shutdown")
async def shutdown_event():
    """Shutdown event."""
    logger.info("Shutting down Image Generation Service")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000) 