image-service-requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
Pillow>=9.5.0
numpy>=1.24.0
six>=1.16.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0 